Template: Select GDAL vs. OGR based on content
Given the broad range for raster and vector formats, lists of file
extensions from GDAL must remain incomplete and ambiguous. So we use
the lists only to choose between GDAL and other readers. In the end,
we choose between GdalTemplate and OgrTemplate based on the driver's
ability to read the actual data. This allows us to handle formats
which don't register extensions.
Resolves GH-67 (OziExplorer files as templates).